Extract storage secrets CLI commands
